chore: move asset selection to top-level components directory
